refactor: consolidate GitHub URL parsing and route validation in lib/github/parse.ts
